update MediaV2 codec selection process for cts/mcts split
mediav2 codec selection process had a presumption that
all codecs would be tested within a single run. This is no longer true
in a cts/mcts separated world.
Update the selection algorithm so that it can insist that a codec
exists for the configuration, but not that it will be tested in this
run. This works in that cts and mcts will both know that there is
a codec and outside of that we know that 1 of them will test it.

Bug: 343583182
Bug: 343310819
Bug: 331566031
Test: atest CtsMediaV2TestCases:android.mediav2.cts.CodecDecoderTest#testSimpleDecodeQueueCSDNative
Test: atest MctsMediaV2TestCases:android.mediav2.cts.CodecDecoderTest#testSimpleDecodeQueueCSDNative
Flag: TEST_ONLY
Merged-In: I39da5f60eb8f0b770c15501d9573e3bdd6a4180a
Change-Id: I39da5f60eb8f0b770c15501d9573e3bdd6a4180a
(cherry picked from commit a5ab8680eaec00000f4584e9d4f9454946eb22a0)