Add MD support for SettingsSuggestionTests
secondary_user_on_secondary_display is for background users that have
access to UI on assigned displays (a.k.a. visible background users) on
devices that have config_multiuserVisibleBackgroundUsers enabled. The
main use case is Automotive's multi-display Whole Cabin experience where
passengers (modeled as visible background users) can interact with the
display in front of them concurrently with the driver (modeled as the
current user) interacting with the driver's display.

Bug: 342492404
Test: atest CtsSettingsSuggestionsTest && atest --user-type secondary_user_on_secondary_display CtsSettingsSuggestionsTest
Change-Id: Ic746bc63922fd415a2ad4278fa9f2bb396c7f066