Return MotionEvent from remote view
Before this CL, the remote view inside
CrossProcessSurfaceControlViewHostTestService was keeping the
MotionEvents to itself, and did some light assertions in this.

This logic is difficult to read and is not very flexible.

Instead, in this CL, just return the MotionEvent that the view
receives, leaving it up to the test to make assertions on it with the
existing matcher infrastructure.

Bug: 329439551
Test: atest android.view.surfacecontrol.cts.SurfaceControlViewHostTests
Change-Id: I0ec1ea4c500b4c77cdf9241c734c2672374994b0