Wait for home activity to be visible during test setUp
Some `ManifestLayoutTests` were found to be flaky when ran as part of
the whole test class but pass fine when ran individually. This is likely
due to the fact that we do not wait for the home activity to become
visible during the setUp and tearDown between tests. As desktop
windowing sets the taskbar to always visible whereas the default home
state has the taskbar as auto hides, its likely this is interfering
with the inset values received throughout the test.

Fixes: 343922712
Test: atest CtsWindowManagerDeviceAnimations:ManifestLayoutTests
Change-Id: I194666beb74891db789e2b85a6079f1e1a504dfa