Restore the logic of verifying the size of DisplayMetrics
The widthPixels and heightPixels of a DisplayMetrics should be reduced
by display cutout and navigation bar insets if it is obtained from an
app which is running on a platform where
Flags.insetsDecoupledConfiguration() is not enabled. This CL restores
the logic.

This CL adapts to the commit: 89d30ec2931b16128e603c8d6579d3503cc02575

Bug: 340970172
Test: atest WindowMetricsActivityTests WindowMetricsWindowContextTests
Change-Id: I1d21c1c53c6378c2516f33471c2f78f3e148297e