Verify behavior for invalid transfer parameters
When creating the ColorSpace for an image, creating the
TransferParameters constructor can fail and throw an Exception. Verify
that the Exception is handled - for this particular image, it means that
the image is successfully decoded and it does *not* have a ColorSpace,
since the encoded color space did not fit the requirements of ColorSpace.

Bug: 198155681
Test: This
Change-Id: I805aba8c078aac71d92ea71a199c2e1fe52df8fd