CtsNfcResolverDeviceTest assumes at most 2 targets
We exercise three types of targets (app, app-activity, and app-filter)
but only display two at a time in our UI, so the test was failing to
find one of the three targets.

This CL breaks the existing test into two similar cases that exercise
configurations with two targets each -- one for app + app-activity,
and one for app + app-filter.

Test: presubmits (w/ modified CTS test)
Bug: 345112999
Flag: EXEMPT bugfix
Change-Id: Ie9855af54ce391f041e2f035edf3ce505f517e07