Stop relying on "secondaryUser" in CtsInputMethodInstallTestCases
This is a follow up CL to my previous CL [1], which rewrote

  CtsInputMethodInstallTestCases

by using bedstead as part of Bug 323240011.

What went wrong was that a "seconary user" can also be the initial
user in bedstead, while the original tests used that term just to mean
some user other than the current user. In bedstead "addtional user" is
a more suitable concept for such test scenarios.

We actually have been using "addtional user" for automotives [2]
because of the same reasons as discussed in Bug 295475990, but there
is no reason to do so only in automotive devices. The idea should work
for all the form factors.

 [1]: I14dc1368d1a26b17c81a616a62431cf0b2c85de4
      019694062286870ef6c59c19d375049105c2b91e
 [2]: I0155b22dad70bf32d423c9a9ce5d72052c32a754
      f8a7a5851813c50f17d99672f3cc5b42ca3e31a8

Fix: 324666126
Test: atest CtsInputMethodInstallTestCases
Test: atest CtsInputMethodInstallTestCases --instant
Change-Id: I71cc9804e4d3e1d8df8a6a815074311047c1c5ff