Moving hardware check from @BeforeClass to @Before.
Because of the check being in @BeforeClass init() the tests are not
running on Watches, TVs, Auto, resulting in the Instrumentation failure
where the tests registered are more than the tests that run.
Moving the check to @Before setUp() will run all the tests but return as
assumption failed on the above mentioned device types.

Bug: 337152039
Test: atest CtsScopedStorageDeviceOnlyTest
Flag: EXEMPT bugfix
Change-Id: I85df67e159afbeb3c9016bc1e797be0af2115615