Camera ITS: Choose a common YUV + JPEG size for test_yuv_jpeg_capture_sameness
We used the size selected for yuv capture, assuming jpeg capture would be
supported as well.

For first  api level < ANDROID15, if no jpeg size is available which
matches YUV, skip the test.

For first api level >= ANDORID15, find the jpeg size that matches the
aspect ratio of YUV (if the same size isn't available), downscale the JPEG
and then compare.

Increase the diff threshold to 3% to account for artifacts created by
cv2 downscale.

Bug: 346930337

Test: test_yuv_jpeg_capture_sameness on device which has different subsets of
      supported jpeg and yuv sizes

Change-Id: I718a41502d6952371f7d35ffbcf5af4f8a216e81
Signed-off-by: Jayant Chowdhary <jchowdhary@google.com>