Allow 64-bit-only devices with less than 2GiB.
The original intention of CDD was that devices with less than 2GiB RAM must support only one ABI. It just happened to be the case at the time that no-one was trying to ship 64-only, so the test was written to assume that meant 32-only.

Fix the test to allow either 32-only or 64-only.

Also rename the test from test32BitOnlyIfLessThan2GiB to testDevicesWithLessThan2GiBMustOnlySupportOneAbi for clarity.

Bug: 331900627
Test: run CtsHardwareTestCases on a 64-only 2GiB device
Signed-off-by: Bian Jin chen <kenjc.bian@rock-chips.com>
Change-Id: Ie4410aed7d6c856bf69a8205fb8a92cfa7895316