Switch back to the initial user in the loop.
After switching to user 10, the test did not switch back to the original
user, so the second iteration was running as user 10.

Switching back to the initial user inside the test is the intended
behavior for the test, and also avoids the timeout error happening
during the tearDown of the bedstead infra.

Bug: 341045459
Flag: TEST_ONLY
Test: atest android.camera.cts.headlesssystemusertest.Camera2HeadlessSystemUserTest

Change-Id: I07727efd08db9eca9789d1fd11900b7b008b88ef