Set the preconditions instead of only failing on them
The hidden API killswitch tests were not working properly because the
settings were not being set correctly. This CL fixes the issue by setting
the settings in the @Before method instead of in a target preparer.

This CL also removes the target preparer that was being used to set the
settings.

Change-Id: Iad343c8a052f5e51b3e25db3d63f06f54d7619e1
Test: atest android.signature.cts.api.current.SignatureTest