Fix BlurTests assumptions about insets
It is not enough to inset the window frame and assume that leaves the
window contents centred. A device can have different insets on different
sides and for this screenshot test we must keep track of where the
actual centre of the window is.

Instead of insetting the window bounds, leave the window bounds at their
full extents but only check pixels inside the insetted window bounds.

Test: atest CtsWindowManagerDeviceAnimations
Bug: 347853456
Flag: EXEMPT bugfix to test
Change-Id: I9dc9d00dfde610da34ef38be0ab49beb5db2b266