Remove message checks in SQLiteCantOpenDatabaseException
The CTS requirement for SQLiteCantOpenDatabaseException is that the
exception is thrown if the database cannot be opened.  The CTS test
should not test for specific sub-strings in the exception, as these
are not part of our API contract.

The substring check is retained as a unit test, however.

Test: atest
 * CtsDatabaseTestCases

Bug: 338297108
Flag: EXEMPT bugfix
Change-Id: I66c40fb295c33e43cde8bf8428625ae24a7bb06a