Add end-to-end tests for certificate blocklist
These tests ensure that an update to the blocklist will effectively be
picked up by an app. These complements the current existing test
(CertBlocklistTest) which checks the built-in blocklists. It supersedes
com.android.server.CertBlacklisterTest which will be removed in a
following change.

Since the loading of the blocklist occurs when an app starts, add a test
helper app (CertBlocklistTestApp). This app exposes an AIDL service
to instruct it to validate a certificate (isBlocked).

A test certificate (blocklist_file_test_ca.pem) and key
(blocklist_file_test_key.pem) are added.

Bug: 340363351
Test: atest CtsSecurityTestCases:android.security.cts.CertBlocklistFileTest
Change-Id: I538888cc7c722b2e1fab2c0d15bf7a9e641572b0