[CTS-V]Set fitsSystemWindows=false for Wear devices
Wear CTS-V builds were mostly practically broken with ag/27388454,
which adds fitsSystemWindows=true to all screen.

The small screen of Wear devices did not work well when the value of the
property fitsSystemWindows was true. This CL changes the value to false
only for watch builds of CTS-V.

The CL works by creating a new boolean config, fit_system_windows,
that is set to "true" by default, but set to "false" for watch
builds only.

Bug: 350828333
Test: manual
Flag: TEST_ONLY
Change-Id: I068d9d3c61e07c149be8fbece58efd88bfb32508