Add CTS for display category task restrictions.
This change adds CTS tests to verify that activities with different
display categories are not placed in the same task. The tests cover
the following scenarios:
1. Launching an activity without a display category when an existing
   task has an activity with a specific display category.
2. Launching activities with different display categories.
3. Launching activities with the same display category.

These tests ensure that the behavior implemented in ag/20257057 is
working as expected, preventing activities with different display
categories from being placed in the same task.

Test: atest CtsVirtualDevicesAppLaunchTestCases:RestrictActivityTaskTest
Fix: 263248504
Flag: TEST_ONLY
Change-Id: I585f310aa6f7aa22880e534b39d84c5e3601af94