Camera: Wait for image reader close to complete
The zoom capture test module default reader is currently
released within the same handler as the image reader callback.
This ensures that the callback is not going to potentially access
invalid frames. Normally releasing the reader in a separate thread
is not an issue because most tests run the sequence only once.
However if we have two or more back to back runs, then releasing
the reader in a separate thread can invalidate it unexpectedly
during the second or any subsequent execution.
To avoid this, ensure that the default reader is always
closed before the test sequence returns back.

Bug: 350877069
Test: atest -c -d cts/tests/camera/src/android/hardware/camera2/cts/ZoomCaptureTest.java --iteration=10

Change-Id: I509b984d355dde9ba68879085d264aa3dfa81cc8