Fix CarDisplayInputLockTest fail with crashes
In the multi-user, multi-display environment that supports background
users with UI access on assigned displays (a.k.a. visible background
users), it's not allowed to create a window or launch an activity on
an unassigned display.
The crash occurs when the test creates an activity on an unassigned
display to check whether the input events are delivered to the
activity or not. Therefore, the test has to check the behavior of
DisplayInputLock with a display assigned to the user running the test.
As a result, the test is rewritten as follows:
- For the current user, it checks if DisplayInputLock ignores
  the attempt to lock the driver display assigned to the current user.
- For the visible background user, it verifies that DisplayInputLock
  successfully locks the passenger display assigned to the visible
  background user.

Bug: 335929246
Flag: EXEMPT test fix
Test: atest --user-type secondary_user_on_secondary_display CtsCarTestCases:CarDisplayInputLockTest
Test: atest CtsCarTestCases:CarDisplayInputLockTest

Signed-off-by: Dohyun Lee <dohyun.lee@lge.com>
(cherry picked from https://partner-android-review.googlesource.com/q/commit:f03859cef8ec0e873f2b0c8d3be74835e8915ae7)
Merged-In: Ibff585bd2361ac8420f57e58c37dd6f4f148d385
Change-Id: Ibff585bd2361ac8420f57e58c37dd6f4f148d385