Enable secondary_user_on_secondary_display in CtsWindowManagerSdk29TestCases
secondary_user_on_secondary_display is for background users that have
access to UI on assigned displays (a.k.a. visible background users) on devices that have config_multiuserVisibleBackgroundUsers enabled. The main use case is Automotive's multi-display Whole Cabin experience where
passengers (modeled as visible background users) can interact with the display in front of them concurrently with the driver (modeled as the the current user) interacting with driver's display.

Bug: 350564574
Test: atest CtsWindowManagerSdk29TestCases
      atest --user-type secondary_user_on_secondary_display
      CtsWindowManagerSdk29TestCases
Flag: TEST_ONLY
Change-Id: I9fe1ec5325d7473eb177edfcc3c9b769d110fdf7