[MockModem] Fix execption as querying network parameters
As SIM card is changed, MockNetworkService would map to a pre-load
network configuration. But currently, only 3 networks are supported. If
the inserted SIM did not belogn to them, a null MockModemCell would be
returned. Add checking if the the MockModemCell is null before querying
the network parameters.

Bug: 343814803
Test: Run CellBroadcastUiTest#testAlertUiOnReceivedAlert
Flag: EXEMPT bugfix

Change-Id: I177d823f247c1b7bf9d85e649b27d3fadf515c95