Enable secondary_user_on_secondary_display for CtsNativeMediaMetricsTestCases
--secondary_user_on_secondary_display is for background users that have
UI access on assigned displays, a.k.a. visible background users, on
devices that have config_multiuserVisibleBackgroundUsers enabled.

The main use case is Automotive's multi-display where passengers
(modeled as visible background users) can interact with the display
in front of them concurrently with the driver (modeled as the the
current user) who interacts with driver's display.

Bug: 353762417

Test: atest --user-type secondary_user_on_secondary_display \
          CtsNativeMediaMetricsTestCases
Test: atest CtsNativeMediaMetricsTestCases
Change-Id: I43ee7bb258ac8f566c43c6b8dce56e789277ecec