Enable secondary_user_on_secondary_display for CtsNativeMediaSlTestCases
--secondary_user_on_secondary_display is for background users that have
UI access on assigned displays, a.k.a. visible background users, on
devices that have config_multiuserVisibleBackgroundUsers enabled.

The main use case is Automotive's multi-display where passengers
(modeled as visible background users) can interact with the display
in front of them concurrently with the driver (modeled as the the
current user) who interacts with driver's display.

Bug: 353762460

Test: atest --user-type secondary_user_on_secondary_display \
          CtsNativeMediaSlTestCases
Test: atest CtsNativeMediaSlTestCases
Change-Id: I4552ab93926bc5aa45886f43c638e9f95f3081fa