Enable secondary_user_on_secondary_display for CtsSystemUiTestCases
secondary_user_on_secondary_display is for background users that have
access to UI on assigned displays (a.k.a. visible background users) on devices that have config_multiuserVisibleBackgroundUsers enabled. The main use case is Automotive's multi-display Whole Cabin experience where
passengers (modeled as visible background users) can interact with the display in front of them concurrently with the driver (modeled as the the current user) interacting with driver's display.

Fixes: 348889913
Test: atest --user-type secondary_user_on_secondary_display CtsSystemUiTestCases
Test: atest --user-type secondary_user CtsSystemUiTestCases
Test: atest CtsSystemUiTestCases
Flag: TEST_ONLY

Change-Id: I479cdff7e7672f83c9f3c4b304d6b795a348b632