Enable secondary_user_on_secondary_display for CtsLibcoreWycheproofBCTestCases
--secondary_user_on_secondary_display is for background users that have
UI access on assigned displays, a.k.a. visible background users, on
devices that have config_multiuserVisibleBackgroundUsers enabled.

The main use case is Automotive's multi-display where passengers
(modeled as visible background users) can interact with the display
in front of them concurrently with the driver (modeled as the the
current user) who interacts with driver's display.

Bug: 354264224

Test: atest --user-type secondary_user_on_secondary_display \
          CtsLibcoreWycheproofBCTestCases
Test: atest CtsLibcoreWycheproofBCTestCases
Change-Id: I53e589cdde8e978c615032e497ce36a3d9607f43