Enable secondary_user_on_secondary_display for CtsWindowManagerDeviceAm
secondary_user_on_secondary_display is for background users that have
access to UI on assigned displays (a.k.a. visible background users) on
devices that have config_multiuserVisibleBackgroundUsers enabled.
The main use case is Automotive's multi-display Whole Cabin experience
where passengers (modeled as visible background users) can interact
with the display in front of them concurrently with the driver
(modeled as the the current user) interacting with driver's display.

Fixes: 349704845
Test: atest --user-type secondary_user_on_secondary_display CtsWindowManagerDeviceAm
Test: atest --user-type secondary_user CtsWindowManagerDeviceAm
Test: atest CtsWindowManagerDeviceAm
Flag: TEST_ONLY
Change-Id: I77d3376c7d2f52515dd93be5aa6a2aa04f37553a