Enable secondary_user_on_secondary_display for CtsInstantAppTests
--secondary_user_on_secondary_display is for background users that have
UI access on assigned displays, a.k.a. visible background users, on
devices that have config_multiuserVisibleBackgroundUsers enabled.

The main use case is Automotive's multi-display where passengers
(modeled as visible background users) can interact with the display
in front of them concurrently with the driver (modeled as the the
current user) who interacts with driver's display.

Bug: 354264982

Test: atest --user-type secondary_user_on_secondary_display \
          CtsInstantAppTests
Test: atest CtsInstantAppTests
Change-Id: I62b28d84273070861e89c2a0eafcdc66e0c6a6b9