Enable secondary_user_on_secondary_display for CtsViewReceiveContentTestCases
--secondary_user_on_secondary_display is for background users that have
UI access on assigned displays, a.k.a. visible background users, on
devices that have config_multiuserVisibleBackgroundUsers enabled.

The main use case is Automotive's multi-display where passengers
(modeled as visible background users) can interact with the display
in front of them concurrently with the driver (modeled as the the
current user) who interacts with driver's display.

Bug: 354256576

Test: atest --user-type secondary_user_on_secondary_display \
          CtsViewReceiveContentTestCases
Test: atest CtsViewReceiveContentTestCases
Change-Id: I2dfb7bac7871c943f4550263d1d2d612b4345494