Wait for onDestroy in WallpaperManagerTest
The method WallpaperManagerTestUtil.performChange() is extensively used
in CTS. It only waits for an `onSurfaceCreated` event, and does not
guarantee that onDestroy() was called for the previous engine. This can
cause flakiness in some tests such as testEngineCallbackCounts.

To fix that, check the current WallpaperManager status before running
performChange(), and make sure we wait that onDestroy() was called the
correct number of times before returning from performChange().

The binder calls from getWallpaperInfo() are very short (typically 1ms)
so this shouldn't add any significant overhead.

Flag: EXEMPT CTS bugfix
Bug: 353631798
Test: atest WallpaperManagerTest --iterations 10
Change-Id: I000b8c402b7b3b4e51b10daa9b09dba66dc507e6