Adds CTS tests for verifying a process can cancel a request it made
Adds a PiP test for dual display mode to verify that the feature
can be cancelled while the requesting app is in Picture in Picture mode.

Also adds a test for DeviceStateManager to verify that a process can
cancel a device state request it made, even if the process is
backgrounded.
Bug: 350930205
Bug: 354772125
Test: DeviceStateManagerTests
Test: ExtensionRearDisplayPresentationTest
Flag: android.hardware.devicestate.feature.flags.device_state_requester_cancel_state
Change-Id: I8c19bf405eae658bd352a30175ca5c56b8e3a8dc

Change-Id: I000fe8003b6081545804b7d3ce13b6585d9d7ed6