Allow WindowManager.getApplicationLaunchKeyboardShortcuts(int) to be hidden
Hidden methods in API interfaces will prevent App developers from being
implementing that interface. However, not every interface in the API is
intended to be implemented by App developers in which case hidden
methods are fine. Unfortunately, there is nothing in the API signature
files which differentiate between the two so the test maintains an
allow list of methods which can be hidden. This change adds the
`WindowManager.getApplicationLaunchKeyboardShortcuts(int)` method to
that list.

Bug: 354106365
Test: TH
Change-Id: Ib3b87ae9ac5d5bf04f853cac200859a60f0abc6c