Enforce waiting for activity being paused/resumed
In some runs of the test, the screen was too fast turned off and on again, so that the activity was not yet paused / resumed.
This leads to the IME not being shown. Adding an explicit check for this test, to be more stable.

Test: atest android.view.inputmethod.cts.KeyboardVisibilityControlTest#testHideImeAfterBackPressed_ScreenOffOn
Flag: EXEMPT bugfix
Fix: 355006526
Bug: 372269584
Change-Id: Ibca67c6a607dc257086b891de9fa5f8ac2f4bc26
Merged-In: Ibca67c6a607dc257086b891de9fa5f8ac2f4bc26