Implement new RingerTest case to test ringer mode change while MT call is RINGING.
This CL adds a new CUJ tests in a RingerTest that tests the scenario where a device is in vibrate mode and then the user changes the ringer to normal mode while an MT call is in the RINGING state. Initially, the device should only vibrate and not ring audibly but when the device is changed to normal mode, the device should both vibrate and ring audibly.

Test: atest RingerTest#testIncomingCallVibrateModeEnableRinger_VibrateAndRing
Flag: TEST_ONLY
Fixes: 352334683
Change-Id: I05c006f793c6b29f3d0ced24a80a91a2d338a267