[webview] Improve resilience and readability of WebSettingsTest image tests
This change uses postmessage to determine when an image has actually loaded (or has indeed been blocked) and then checks the height of the image via javascript to ensure it has loaded correctly or been blocked.

Context:

WebSettingsTest image loading tests currently use the page title to check image heights to determine if images have loaded or been blocked correctly. This is done by setting the page title to the height of the image in the image's onload function and then querying the page title.

For some tests a polling check is used to wait for the page title to change and indicate that the image has loaded. This is not very efficient. For other tests it simply checks right after the page finishes loading. This approach may fail if the image is delayed in loading and the title is checked too soon.

Fixes: b/325215912
Test: atest CtsWebkitTestCases:WebSettingsTest
Flag: TEST_ONLY
Change-Id: I8a8a6031f9f88394c856a3d27a267b5260280626