Tap on embedded view instead of surface view
This is a partial revert of aosp/3119597. In that CL, we switched to the new method of injection - to send input to the embedded view instead of to the surface view.

However, this likely required other fixes, and possibly in the platform. It's difficult to identify which specific patches are missing. Therefore, in this CL, we partially revert the previous commit to use the old behaviour, which makes the test pass.

The test is already working fine on main so nothing is needed to be done there.

Bug: 352656601
Flag: TEST_ONLY
Test: USE_RBE=false atest CtsWindowManagerDeviceTestCases:android.server.wm.SurfaceControlViewHostTests#testCanReplaceSurfacePackage
(cherry picked from https://android-review.googlesource.com/q/commit:c0af3053e885c3d0ec74e9b585c38485faa62a30)
Merged-In: I07ff0a88d66fd70b1e1dd22c54a3d2b954733cc5
Change-Id: I07ff0a88d66fd70b1e1dd22c54a3d2b954733cc5