Use toolType that's right for the emulated device
When uinput devices are created, we typically expect a certain tooltype
to be produced. For example, touchscreen should produce MT_TOOL_FINGER
and the stylus should produce MT_TOOL_PEN.

Update the UinputTouchDevice to require this tooltype to be specified,
and then update the current usages to conform to this requirement.

This will allow us to more accurately represent physical devices.

Bug: 329439551
Test: atest CtsInputTestCases
Flag: TEST_ONLY
Change-Id: Ib96a5ef1e87daaee895dc5309b4f17e9c7f91815