Scroll to view in IME Switcher menu tests
If the display size is small, or there are many enabled IME subtypes
when the test is ran, it may fail to find the item corresponding to the
test IME, or the language settings button. This ensures we scroll
through the list to try to find the items as needed.

This also takes into account cases when the list is not scrollable.

The ImeSettings used for testing is modified to support suppressing
setting the test IME as the current one, but rather only initialize and
enabled it, to avoid errors being logged, as the IME is selected without
being properly initialized, in testInputMethodPickerSwitchIme.

Flag: android.view.inputmethod.ime_switcher_revamp
Test: atest testShowInputMethodPicker testInputMethodPickerSwitchIme
  testInputMethodPickerOpenLanguageSettings
Bug: 311791923
Change-Id: I36faf38ce4f0f3c40c68e9c684393a8f1b3dda39