Log SecureSettings in getCurrentInputMethodId() (2nd try)
This CL introduces some additional debug info into

  MockImeSession#getCurrentInputMethodId()

to diagnose further on why

  InputMethodManager#getCurrentInputMethodInfo()

does not start returning the IME ID for MockIme after

  adb shell ime enable --user <userId> <MockImeId>
  adb shell ime set --user <userId> <MockImeId>

This CL only adds new logs to logcat. There must be no change in the
test expectations.

Note that why the the previous attempt [1] broke

  CtsAutoFillServiceTestCases

was because Settings.Secure.getString() is already locked down with
targetSdkVersion. We could make CtsAutoFillServiceTestCases.apk
testOnly to work around the limitation, requesting all the clients of
MockIme to specify "testOnly" probably does not scale.

This version uses shell command

  adb shell settings get --user <userId> secure <settingsId>,

which is supposed to be more robust.

 [1]: Icc481341988583aa7ccb39052ae31e31277a2841
      39c8763a3c551699c42b4bbb36e3f526cb804b40

Bug: 354782333
Flag: TEST_ONLY
Test: atest CtsInputMethodTestCases
Test: atest CtsAutoFillServiceTestCases
Change-Id: I6d828856629cfccc34d1769c21880e988c6382aa