Update test to wait for keyguard gone to increase resilence to timing issues.
The keyguard_wm_state_refactor flag ties the keyguard going away transition to a SpringAnimation, making it nondeterministic. The flag also moves control of the keyguard showing/going away logic to flows, which changes the timing of these state changes. This test is flaky with the flag enabled due to these timing issues, but passes consistently if we simply wait for the state to settle.

Bug: 278086361
Flag: TEST_ONLY
Test: atest KeyguardTests
Change-Id: Idf50ac5da9ca4d5100df94aaea7022298c579cd0