Don't add empty list to received frames
Depending on how quickly the service binds, we may get two calls to
notifyPollingFrames. One with all the frames, and a second with an empty
list, because HostEmulationManager sends pending frames on service
connection. It seems reasonable for an app to just ignore an empty list
of frames, so our test service should do the same.

Bug: 343572313
Test: CtsNfcTestCases:CardEmulationTest
Change-Id: I7283000eccd8b00b99cab2d9592c19ef094c5afc