Fix NotificationManagerZenTest policy comparisons before MODES_UI
With MODES_API but without MODES_UI, the default policy (for the purposes of filling out automatic zen rules created with a missing or underspecified zen policy) is not the device-default policy, but rather the manual DND policy at the time of rule creation. Therefore, calculate mDefaultPolicy to match the previous call to setNotificationPolicy() in the test setup.

Bug: 357126912
Test: atest NotificationManagerZenTest, after thoroughly changing default_zen_mode_config.
Flag: EXEMPT fixing CTS
Change-Id: Ifefa318c4c383182e44a7695d2f4d59c8fa004b8