Fix testEmbeddedViewReceivesRawInputCoordinatesInDisplaySpace
This test was failing on devices where physical and logical display spaces are different. Previously, the test sent the touch in logical display space when it should be sent in physical display. This CL fixes the issue by mapping the touch coordinates from logical display space to physical display space.

Bug: 357552749
Flag: EXEMPT bugfix
Test: ran test with different display sizes via `adb shell wm size ...`
Change-Id: I2a597d1bcbb988a80a7bb49ce7814bb596a42f05