Improve logging for visual voicemail CTS Verifier Test
Improved the logging in the VVM CTS verifier test.  It was previously
impossible to correlate logs to test steps.  These additional logs make
it possible to determine which test step failed and to then be able to
look at Telephony logs for more insight into what happened.

Test: This itself is a test; verified logging in logcat is present.
Flag: EXEMPT bugfix
Bug: 354090891
Change-Id: Ic13d8ecc70c105f8b5b93f3f6dca9a67888d0835