Provide epsilon when comparing coords in InputEventMatchers
When the injection is done via "injectInputEvent" or other means, we are
able to maintain high precision. When the injection is done via
UinputTouchDevice, we are losing some precision due to transforms. Since
the matchers have no way of knowing how the event was injected, we must
force the caller to provide the tolerance information.

In this CL, we are forcing the java users of coord comparisons to
provide the tolerances, and the kotlin users will remain the same
because we will keep the old behavior by default.

This CL allows tests like "WallpaperWindowInputTests" to pass.
Previously, this test was failing locally. It was also failing when the
"wm size" command was executed to change the logical display size.

Bug: 357552752
Flag: TEST_ONLY
Test: atest CtsWindowManagerDeviceInput:WallpaperWindowInputTests
Test: adb shell wm size 600x400 (also run this command prior to running
the test above)

Change-Id: Iddea92cedf2edbeb9e50cb868eb19a26aea40068