CtsPrintTestCases: Tag subset for presubmit tests
Running the whole CtsPrintTestCases module takes too long for a
presubmit test.  In order to enable printing on cuttlefish, tag a subset
that is able to finish before the deadline.  The selected subset
provides some coverage on all the publicly documented classes, but moves
the more thorough tests to postsubmit.

Bug: 356221362
Test: atest -p cts/tests/tests/print:presubmit
Flag: TEST_ONLY
Change-Id: If427acef6afd21e32b090194556db428222a6551