Deflake WebViewSslTest#testClearSslPreferences
WebViewSslTest#testClearSslPreferences was flaking occasionally as the
SSL error callback was not being invoked at the expected time.

WebView requests a favicon from sites after a page loads successfully,
which causes an additional request to be sent to the same host that the
test was not expecting. If this request comes after the test has cleared
the cached SSL preferences, it can disrupt the test by creating a new
cached entry sooner than expected.

Avoid this by ignoring the favicon request in the SSL error handler:
don't consider it a successful callback invocation for the purposes of
the test assertions, and cancel the request rather than proceeding past
the error. Cancel decisions are not cached (unlike proceed), so this
should not prevent the expected callback from happening.

Bug: 353493560
Test: atest WebViewSslTest
Flag: TEST_ONLY
Change-Id: Ifb0689040129695a3f99eaa0c314fcdc8f17081b