RESTRICT AUTOMERGE: Add assumption about windowing mode in AspectRatioTestsBase.
Assume that the windowing mode requested for the test is the one that the activity actually launched in when checking for aspect ratio. The relevant tests all check for WINDOWING_MODE_FULLSCREEN, which seems to be explcitly requested.

Note that this test has been deleted for Android 15, so restricting
automerge.

Bug: 349189250
Test: atest AspectRatioTests
Flag: TEST_ONLY
Change-Id: Id8c251a26cf665e673fd957dae8efaca50628fff