Don't spawn sharesheet for tests which verify */* mime type
The ActionGetContentOnlyTest has 2 tests which verify that DocumentsUI
is spawned when an ACTION_GET_CONTENT intent is used with a wildcard
*/* mime type. Unfortunately the tests wrap the original intent with
one which forces the sharesheet open, this causes the tests to not
verify the behaviour on DocumentsUI and instead hang as the sharesheet
is showing instead. Remove the intent wrapping and prefer the original
intent instead.

Bug: 359671894
Test: atest ActionGetContentOnlyTest
Flag: EXEMPT test only change
Change-Id: If8979ef7a0de4e8cae5c35ae93f0d125df97419c