Fix CtsWindowManagerDeviceOther:android.server.wm.other.DragDropCompatTest and CtsWindowManagerDeviceOther:android.server.wm.other.DragDropTest for the visible background user
When triggering the MotionEvent, the displayId value is always set to
DEFAULT_DISPLAY, causing the test to fail for the visible background
user logged into another display.
I have modified it so that when a MotionEvent is triggered,
the displayId value of the user being tested is used, allowing tests
to be performed for the visible background user as well.

Bug: 356745195
Test: atest CtsWindowManagerDeviceOther:android.server.wm.other.DragDropCompatTest
Test: atest --user-type secondary_user CtsWindowManagerDeviceOther:android.server.wm.other.DragDropCompatTest
Test: atest --user-type secondary_user_on_secondary_display CtsWindowManagerDeviceOther:android.server.wm.other.DragDropCompatTest
Test: atest CtsWindowManagerDeviceOther:android.server.wm.other.DragDropTest
Test: atest --user-type secondary_user CtsWindowManagerDeviceOther:android.server.wm.other.DragDropTest
Test: atest --user-type secondary_user_on_secondary_display CtsWindowManagerDeviceOther:android.server.wm.other.DragDropTest
(cherry picked from https://partner-android-review.googlesource.com/q/commit:bf36091aaf4a4cc011c7754314fe3d6117667170)

Change-Id: I8332a3c210eb3c5c6ad00d4bf34c5230b1f4d47c