Handle devices without WebView in WebViewUpdateManagerTest.
WebViewUpdateService doesn't run on devices without a WebView, so
fetching a WebViewUpdateManager triggers errors and WTFs. The API has
been updated to require that it's only used on devices with
FEATURE_WEBVIEW to avoid this.

Instead of having the test expect valid API results in this case (which
did not actually work), just skip the tests on devices without a WebView
to match the updated API requirements.

Bug: 319292658
Test: atest WebViewUpdateManagerTest
Flag: android.webkit.update_service_ipc_wrapper
Change-Id: Ic695992e47fce3a7faf8dd74068a1918a12f3683