HdrToSdrDecoderTest: Pass profile as an argument
Certain vendor extractor implementation may not return profile for some
VP9 clips in older versions of Android. Since this test is part of MTS,
and new expectations can't be imposed on vendor extractors in older
versions of Android, the test is updated to pass profile as argument.

If the extractor returns profile, then it is expected to match this new
argument.

ExtractorTest$ValidateKeyValuePairs in CtsMediaV2TestCases ensures that
the extractors signal profile for these cases.

Bug: 361209021
Test: atest CtsMediaDecoderTestCases:HdrToSdrDecoderTest
Test: atest MctsMediaDecoderTestCases:HdrToSdrDecoderTest

Change-Id: I62dc7c8c149b2e747ce6c20272a0e5203c0019f5