Reduce logging in MediaPlayerRandomTest
Instead of individually logging every action (there are 100,000 of
them!), we instead log a summary and the last 20 actions in detail at
the end of the test (whether it was a success or failure).

It was found the removing the logging completely resolved a test timeout
on Wear OS devices: b/358308326#comment9. This change is designed to fix
the timeout, while retaining most of the usefulness of the logging.

Test: atest MediaPlayerRandomTest
Bug: 358308326
Change-Id: I65acda318847f7d6d1d653d378177d943e7ec19c