Fix testVirtualDisplayWithoutCutout flakiness.
The safe insets are always computed by WM and overridden in
DisplayInfo. The flakiness in the test is caused by the race
between the WM override coming in and the test expectation.

We just set the safe insets to what WM will compute anyway
because the purpose of this test is not to test the WM cutout
override logic.

Fix: 355646121
Fix: 360152105
Test: atest
Flag: EXEMPT test-only change
Change-Id: If2eeb36469ae7a9601fcffbabcd4e9b77609bae1