Allow VirtualDisplayActivityScenario to be used as an AutoClosable
We want to be able to run the VirtualDisplayActivityScenario with a test
function in some cases, rather than only as a test rule. To do this, we
extract a new interface for the scenario, and create two concrete
implementations: one as a test rule and one as an AutoClosable.

Bug: 310997010
Test: Presubmit
Flag: TEST_ONLY
Change-Id: Ic1b00b4fd38b9fd8d51f17da00979751083bcfc7