Fix tearDown in ActivityEmbeddingTestBase
The setUp method of this class has an assumption that activity
embedding is supported, and the test case is skipped for devices
that has no such support (assumeActivityEmbeddingSupportedDevice).
In that case, none of the class fields get initialized, including
the mReportedDisplayMetrics field, making the fields null. So, in
the tearDown method (which gets called regardless of the device's
support for activity embedding), we need to check that the
mReportedDisplayMetrics field is not null before calling any method
on it.

Bug: 365360800
Change-Id: I8be78b2121b8f20b735a12e3b4c89600e0961342
Flag: TEST_ONLY