Hide navigation bar protection during blur tests
By changing the parent theme for Blur BackgroundActivity to
a Material-based theme which sets the bar colors to transparent
and declares windowDrawsSystemBarBackgrounds.

This prevents a black bar at the bottom of the screen from
corrupting some of the edge pixel values after recent changes
to not use immersive mode. There is still a theoretical
interference from navigation bar button pixels in the screenshots
but it is much smaller than with a full rectangle of black pixels.

Bug: 364538048
Change-Id: Ic829c1f80cb9c14f9fa3fbcf86dbf7c764ec12a9
Test: atest CtsWindowManagerDeviceAnimations:BlurTests