Fix clearing accounts before flaky test
Due to unpredictable order of DeviceOwnerTest suite run, sometimes there was an account already on the device when starting DeviceOwnerTest#setDeviceOwner_receivesOwnerChangedBroadcast test. This led to unsuccessful setting of the device owner and test fail.

This commit adds annotations which will clear any existing account before running the flaky test - or skip it, if the operation proves to be impossible to perform.

Bug: b/347418954
Flag: TEST_ONLY
Change-Id: I520e8f41eeae6d261026798bd0674307c9999ddf