Enable secondary_user_on_secondary_display for CtsMediaCUJLargeTest
secondary_user_on_secondary_display is for background users that have
access to UI on assigned displays (a.k.a. visible background users) on
devices that have config_multiuserVisibleBackgroundUsers enabled. The
main use case is Automotive's multi-display Whole Cabin experience where
passengers (modeled as visible background users) can interact with the
display in front of them concurrently with the driver (modeled as the
current user) interacting with driver's display.

Bug: 342431015

Test: atest CtsMediaCUJLargeTest
Test: atest --user-type secondary_user_on_secondary_display \
        CtsMediaCUJLargeTest
Change-Id: Ide55fa2d5c0cf22e47dfd680e09c81eb44054979