Enable secondary_user_on_secondary_display in CtsCompanionDeviceManagerNoCompanionServicesTestCases
secondary_user_on_secondary_display is for background users that have
access to UI on assigned displays (a.k.a. visible background users) on devices that have config_multiuserVisibleBackgroundUsers enabled. The main use case is Automotive's multi-display Whole Cabin experience where
passengers (modeled as visible background users) can interact with the display in front of them concurrently with the driver (modeled as the the current user) interacting with driver's display.

Bug: 364413990
Test: atest CtsCompanionDeviceManagerNoCompanionServicesTestCases
      atest --user-type secondary_user_on_secondary_display CtsCompanionDeviceManagerNoCompanionServicesTestCases
Flag: TEST_ONLY
Change-Id: Ibedcd49fd900cc5a709fc844ecacc4d41bd7dc92