Enable secondary_user_on_secondary_display on CtsPackageSchemeTests
secondary_user_on_secondary_display is for background users that have
access to UI on assigned displays (a.k.a. visible background users) on devices
that have config_multiuserVisibleBackgroundUsers enabled. The main use
case is Automotive's multi-display Whole Cabin experience where
passengers (modeled as visible background users) can interact with the display
in front of them concurrently with the driver (modeled as the the
current user) interacting with driver's display.

Test: atest -c CtsPackageSchemeTestsWithoutVisibility
Test: atest -c --user-type secondary_user_on_secondary_display CtsPackageSchemeTestsWithoutVisibility
Test: atest -c CtsPackageSchemeTestsWithVisibility
Test: atest -c --user-type secondary_user_on_secondary_display CtsPackageSchemeTestsWithVisibility
Bug: 364413990
Flag: TEST_ONLY
Change-Id: I62e4ce46bfb78326db9fda19d2f5293a3ddc5ba4