Enable secondary_user_on_secondary_display on DownloadManagerInstallerTest
secondary_user_on_secondary_display is for background users that have
access to UI on assigned displays (a.k.a. visible background users) on devices
that have config_multiuserVisibleBackgroundUsers enabled. The main use
case is Automotive's multi-display Whole Cabin experience where
passengers (modeled as visible background users) can interact with the display
in front of them concurrently with the driver (modeled as the the
current user) interacting with driver's display.

Test: atest -c DownloadManagerInstallerTest
Test: atest -c --user-type secondary_user_on_secondary_display DownloadManagerInstallerTest
Bug: 364413990
Flag: TEST_ONLY
Change-Id: Icc7ae161c254a9b65b9a60a64823b1b763a475e3