Enable null GameManager on Wear devices
Wear devices can now choose to not start the GameManagerService, which
results in an absent GameManager. These changes skip tests that depend
on the service for Wear devices that do not publish a GameManager.
However, if a Wear device publishes a GameManager, it will still be held
to the same CTS standard as any other device that publish a GameManager.

Bug: 340929737
Test: atest CtsGameManagerTestCases CtsGameFrameRateTestCases
Flag: TEST_ONLY
Change-Id: Ie881c98e055d8a629ec9273d034ea59b32f7f894