Enable secondary_user_on_secondary_display for CtsRsCppTestCases
secondary_user_on_secondary_display is for background users that have
access to UI on assigned displays (a.k.a. visible background users) on
devices that have config_multiuserVisibleBackgroundUsers enabled.
The main use case is Automotive's multi-display Whole Cabin experience
where passengers (modeled as visible background users) can interact
with the display in front of them concurrently with the driver
(modeled as the the current user) interacting with driver's display.

Fixes: 363271625
Test: atest CtsRsCppTestCases
Test: atest --user-type secondary_user CtsRsCppTestCases
Test: atest --user-type secondary_user_on_secondary_display  CtsRsCppTestCases
Flag: TEST_ONLY

Change-Id: Ia8773449c63b580db16b3db31c2c22e290bbb34e