Enable secondary_user_on_secondary_display for CtsVirtualDevicesTestCases
secondary_user_on_secondary_display is for background users that have
access to UI on assigned displays (a.k.a. visible background users) on
devices that have config_multiuserVisibleBackgroundUsers enabled.
The main use case is Automotive's multi-display Whole Cabin experience
where passengers (modeled as visible background users) can interact
with the display in front of them concurrently with the driver
(modeled as the the current user) interacting with driver's display.

Fixes: 351050708
Test: atest  CtsVirtualDevicesTestCases
Test: atest --user-type secondary_user CtsVirtualDevicesTestCases
Test: atest --user-type secondary_user_on_secondary_display CtsVirtualDevicesTestCases
Flag: TEST_ONLY

Change-Id: I4ad54daa072b70024527db2ccfbd44c7f54cb512