Enable secondary_user_on_secondary_display for CtsWindowManagerBackgroundActivityTestCases
secondary_user_on_secondary_display is for background users that have
access to UI on assigned displays (a.k.a. visible background users) on
devices that have config_multiuserVisibleBackgroundUsers enabled.
The main use case is Automotive's multi-display Whole Cabin experience
where passengers (modeled as visible background users) can interact
with the display in front of them concurrently with the driver
(modeled as the the current user) interacting with driver's display.

This CL enabled secondary_user_on_secondary_display for
CtsWindowManagerBackgroundActivityTestCases. In other words, the
entire module can pass when running as
secondary_user_on_secondary_display with this CL.

This CL also fixed some failed
CtsWindowManagerBackgroundActivityTestCases for the driver. They
failed on AAOS multi-user multi-display devices because the test
activity is the top focused activity on its own display but the
display itself might not be the top focused display. The test
asserted the test activity was the top focused activity among all
displays, which was unnecessary.
Two tests are still failing when running as the driver with this CL,
and they're tracked by b/363179542

Fixes: 355320922
Fixes: 351081318
Test: atest CtsWindowManagerBackgroundActivityTestCases (two tests are still failing)
Test: atest --user-type secondary_user CtsWindowManagerBackgroundActivityTestCases
Test: atest --user-type secondary_user_on_secondary_display CtsWindowManagerBackgroundActivityTestCases
Flag: TEST_ONLY

Change-Id: If7e757048b4b4d2fe4bf96fed25c6ef69e236c7a